+++ src/brewspy.cpp
@@ -1,7 +1,7 @@
 /*
 MIT License
 
-Copyright (c) 2021-22 Magnus
+Copyright (c) 2021-2024 Magnus
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -60,7 +60,7 @@ void Brewspy::sendTapInformation(UnitIndex idx, float stableVol,
   doc["pour"] = pourVol;
   Log.notice(F("BSPY: Sending TAP information to brewspy, last %Fl, "
                "pour %Fl  [%d]" CR),
-             stableVol, pourVol);
+             stableVol, pourVol, idx);
   String out;
   out.reserve(100);
   serializeJson(doc, out);
@@ -69,7 +69,11 @@ void Brewspy::sendTapInformation(UnitIndex idx, float stableVol,
   EspSerial.print(out.c_str());
   EspSerial.print(CR);
   // #endif
-  _push->sendHttpPost(out, "https://brew-spy.com/api/tap/keg/set", "", "");
+  out =
+      _push->sendHttpPost(out, "https://brew-spy.com/api/tap/keg/set", "", "");
+  updateStatus(out);
+
+  Log.info(F("BSPY: Response %s." CR), out.c_str());
 }
 
 void Brewspy::sendPourInformation(UnitIndex idx, float pourVol) {
@@ -98,7 +102,11 @@ void Brewspy::sendPourInformation(UnitIndex idx, float pourVol) {
   EspSerial.print(out.c_str());
   EspSerial.print(CR);
   // #endif
-  _push->sendHttpPost(out, "https://brew-spy.com/api/tap/keg/pour", "", "");
+  out =
+      _push->sendHttpPost(out, "https://brew-spy.com/api/tap/keg/pour", "", "");
+  updateStatus(out);
+
+  Log.info(F("BSPY: Response %s." CR), out.c_str());
 }
 
 void Brewspy::clearKegInformation(UnitIndex idx) {
@@ -123,10 +131,14 @@ void Brewspy::clearKegInformation(UnitIndex idx) {
   EspSerial.print(out.c_str());
   EspSerial.print(CR);
   // #endif
-  _push->sendHttpPost(out, "https://brew-spy.com/api/tap/keg/clear", "", "");
+  out = _push->sendHttpPost(out, "https://brew-spy.com/api/tap/keg/clear", "",
+                            "");
+  updateStatus(out);
+
+  Log.info(F("BSPY: Response %s." CR), out.c_str());
 }
 
-String Brewspy::getTapInformation(const String& token) {
+void Brewspy::getTapInformation(JsonObject& obj, const String token) {
   // API: https://brew-spy.com/api/json/taplist/<TOKEN>
   // Descr: To view the current data for a tap you can use the
   // Payload:
@@ -144,7 +156,7 @@ String Brewspy::getTapInformation(const String& token) {
   //     "volUnit" : <volume of beer left unit>,
   //   }
   // }
-  if (token.length() == 0) return "{}";
+  if (token.length() == 0) return;
 
   Log.notice(F("BSPY: Requesting TAP information from brewspy." CR));
 
@@ -152,11 +164,28 @@ String Brewspy::getTapInformation(const String& token) {
   String payload = "";
   String resp =
       _push->sendHttpGet(payload, url.c_str(), "Accept: application/json", "");
-  // #if LOG_LEVEL == 6
   EspSerial.print(resp.c_str());
   EspSerial.print(CR);
-  // #endif
-  return resp;
+
+  DynamicJsonDocument doc(JSON_BUFFER_SIZE_M);
+  DeserializationError error = deserializeJson(doc, resp);
+
+  if (error) {
+    Serial.print(F("deserializeJson() failed: "));
+    Serial.println(error.f_str());
+    return;
+  }
+
+  obj["recipe"] = doc["recipe"].as<String>();
+  obj["abv"] = doc["abv"].as<float>();
+}
+
+void Brewspy::updateStatus(String& response) {
+  _lastTimestamp = millis();
+  _lastStatus = _push->wasLastSuccessful();
+  _lastHttpError = _push->getLastResponseCode();
+  _lastResponse = response;
+  _hasRun = true;
 }
 
 // EOF