+++ src/homeassist.hpp
@@ -1,7 +1,7 @@
 /*
 MIT License
 
-Copyright (c) 2021-22 Magnus
+Copyright (c) 2021-2024 Magnus
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -31,12 +31,25 @@ class HomeAssist {
  protected:
   BasePush *_push;
 
+  bool _hasRun = false;
+  uint32_t _lastTimestamp = 0;
+  bool _lastStatus = 0;
+  int _lastMqttError = 0;
+
+  void updateStatus();
+
  public:
   explicit HomeAssist(BasePush *push) { _push = push; }
 
   void sendTempInformation(float tempC);
   void sendTapInformation(UnitIndex idx, float stableVol, float glasses);
   void sendPourInformation(UnitIndex idx, float pourVol);
+
+  bool hasRun() { return _hasRun; }
+  uint32_t getLastTimeStamp() { return _lastTimestamp; }
+  bool getLastStatus() { return _lastStatus; }
+  int getLastError() { return _lastMqttError; }
+  String getLastResponse() { return ""; }
 };
 
 #endif  // SRC_HOMEASSIST_HPP_