+++ src_docs/source/releases.rst
@@ -3,6 +3,20 @@
 Releases 
 ########
 
+v1.0.0-beta2
+============
+
+* Added new integration to Barhelper
+
+v1.0.0-beta1
+============
+
+* Added influx integration for scale data (debugging)
+* Added favicon
+* Fixed timeout issue when uploading new firmware
+* Created new user interface in VueJS
+* Changed all API's to use JSON
+
 v0.9.0
 ======